requirements.txt
nibabel
numpy
pillow